+++ src/compiler/sema_types.c
@@ -202,12 +202,8 @@ static bool sema_resolve_type_identifier(SemaContext *context, TypeInfo *type_in
 	}
 	Decl *decl = sema_resolve_symbol(context, type_info->unresolved.name, type_info->unresolved.path, type_info->span);
 
-	assert(decl);
 	// Already handled
-	if (!decl_ok(decl))
-	{
-		return type_info_poison(type_info);
-	}
+	if (!decl) return type_info_poison(type_info);
 
 	decl = decl_flatten(decl);
 	switch (decl->decl_kind)